+++ rows/cli.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 
-# Copyright 2014-2017 Álvaro Justen <https://github.com/turicas/rows/>
+# Copyright 2014-2018 Álvaro Justen <https://github.com/turicas/rows/>
 
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Lesser General Public License as published by
@@ -31,11 +31,15 @@ from io import BytesIO
 import click
 import requests.exceptions
 import requests_cache
+from tqdm import tqdm
 
 import rows
+import six
 from rows.plugins.utils import make_header
-from rows.utils import (detect_source, export_to_uri, import_from_source,
-                        import_from_uri)
+from rows.utils import (csv2sqlite, detect_source, export_to_uri,
+                        import_from_source, import_from_uri, pgexport,
+                        pgimport, ProgressBar, sqlite2csv, uncompressed_size)
+
 
 DEFAULT_INPUT_ENCODING = 'utf-8'
 DEFAULT_INPUT_LOCALE = 'C'
@@ -46,11 +50,16 @@ CACHE_PATH = HOME_PATH / '.cache' / 'rows' / 'http'
 
 
 def _import_table(source, encoding, verify_ssl=True, *args, **kwargs):
+    # TODO: add `--quiet|-q` or `--progress|-p` to set `progress` properly
     try:
-        table = import_from_uri(source,
-                                default_encoding=DEFAULT_INPUT_ENCODING,
-                                verify_ssl=verify_ssl,
-                                encoding=encoding, *args, **kwargs)
+        table = import_from_uri(
+            source,
+            default_encoding=DEFAULT_INPUT_ENCODING,
+            verify_ssl=verify_ssl,
+            encoding=encoding,
+            progress=True,
+            *args, **kwargs,
+        )
     except requests.exceptions.SSLError:
         click.echo('ERROR: SSL verification failed! '
                    'Use `--verify-ssl=no` if you want to ignore.', err=True)
@@ -92,6 +101,24 @@ def _get_export_fields(table_field_names, fields_exclude):
     else:
         return None
 
+def _get_schemas_for_inputs(schemas, inputs):
+    if schemas is None:
+        schemas = [None for _ in inputs]
+    else:
+        schemas = [schema.strip() or None for schema in schemas.split(',')]
+        if len(schemas) > len(inputs):
+            click.echo(
+                'ERROR: number of schemas is greater than sources', err=True,
+            )
+            sys.exit(9)
+        elif len(schemas) < len(inputs):
+            diff = len(inputs) - len(schemas)
+            for _ in range(diff):
+                schemas.append(None)
+
+    return [rows.fields.load_schema(schema) if schema else None
+            for schema in schemas]
+
 
 @click.group()
 @click.option('--http-cache', type=bool, default=True)
@@ -110,8 +137,8 @@ def cli(http_cache, http_cache_path):
 
 
 @cli.command(help='Convert table on `source` URI to `destination`')
-@click.option('--input-encoding')
-@click.option('--output-encoding')
+@click.option('--input-encoding', default='utf-8')
+@click.option('--output-encoding', default='utf-8')
 @click.option('--input-locale')
 @click.option('--output-locale')
 @click.option('--verify-ssl', type=bool, default=True)
@@ -157,11 +184,11 @@ def convert(input_encoding, output_encoding, input_locale, output_locale,
 
 @cli.command(help='Join tables from `source` URIs using `key(s)` to group '
                   'rows and save into `destination`')
-@click.option('--input-encoding')
-@click.option('--output-encoding')
+@click.option('--input-encoding', default='utf-8')
+@click.option('--output-encoding', default='utf-8')
 @click.option('--input-locale')
 @click.option('--output-locale')
-@click.option('--verify-ssl', default=True, type=bool)
+@click.option('--verify-ssl', type=bool, default=True)
 @click.option('--order-by')
 @click.option('--fields',
               help='A comma-separated list of fields to export')
@@ -210,11 +237,11 @@ def join(input_encoding, output_encoding, input_locale, output_locale,
 
 @cli.command(name='sum',
              help='Sum tables from `source` URIs and save into `destination`')
-@click.option('--input-encoding')
-@click.option('--output-encoding')
+@click.option('--input-encoding', default='utf-8')
+@click.option('--output-encoding', default='utf-8')
 @click.option('--input-locale')
 @click.option('--output-locale')
-@click.option('--verify-ssl', default=True, type=bool)
+@click.option('--verify-ssl', type=bool, default=True)
 @click.option('--order-by')
 @click.option('--fields',
               help='A comma-separated list of fields to import')
@@ -259,12 +286,14 @@ def sum_(input_encoding, output_encoding, input_locale, output_locale,
 
 
 @cli.command(name='print', help='Print a table')
-@click.option('--input-encoding')
-@click.option('--output-encoding')
+@click.option('--input-encoding', default='utf-8')
+@click.option('--output-encoding', default='utf-8')
 @click.option('--input-locale')
 @click.option('--output-locale')
+@click.option('--frame-style', default='ascii',
+              help='Options: ascii, single, double, none')
 @click.option('--table-index', default=0)
-@click.option('--verify-ssl', default=True, type=bool)
+@click.option('--verify-ssl', type=bool, default=True)
 @click.option('--fields',
               help='A comma-separated list of fields to import')
 @click.option('--fields-exclude',
@@ -272,7 +301,8 @@ def sum_(input_encoding, output_encoding, input_locale, output_locale,
 @click.option('--order-by')
 @click.argument('source', required=True)
 def print_(input_encoding, output_encoding, input_locale, output_locale,
-           table_index, verify_ssl, fields, fields_exclude, order_by, source):
+           frame_style, table_index, verify_ssl, fields, fields_exclude,
+           order_by, source):
 
     import_fields = _get_import_fields(fields, fields_exclude)
     # TODO: if create_table implements `fields_exclude` this _import_table call
@@ -303,10 +333,10 @@ def print_(input_encoding, output_encoding, input_locale, output_locale,
     if output_locale is not None:
         with rows.locale_context(output_locale):
             rows.export_to_txt(table, fobj, encoding=output_encoding,
-                               export_fields=export_fields)
+                               export_fields=export_fields, frame_style=frame_style)
     else:
         rows.export_to_txt(table, fobj, encoding=output_encoding,
-                           export_fields=export_fields)
+                           export_fields=export_fields, frame_style=frame_style)
 
     fobj.seek(0)
     # TODO: may pass unicode to click.echo if output_encoding is not provided
@@ -314,16 +344,22 @@ def print_(input_encoding, output_encoding, input_locale, output_locale,
 
 
 @cli.command(name='query', help='Query a table using SQL')
-@click.option('--input-encoding')
-@click.option('--output-encoding')
+@click.option('--input-encoding', default='utf-8')
+@click.option('--output-encoding', default='utf-8')
 @click.option('--input-locale')
 @click.option('--output-locale')
-@click.option('--verify-ssl', default=True, type=bool)
+@click.option('--verify-ssl', type=bool, default=True)
+@click.option('--samples', type=int, default=5000,
+              help='Number of rows to determine the field types (0 = all)')
 @click.option('--output')
+@click.option('--frame-style', default='ascii',
+              help='Options: ascii, single, double, none')
 @click.argument('query', required=True)
 @click.argument('sources', nargs=-1, required=True)
 def query(input_encoding, output_encoding, input_locale, output_locale,
-          verify_ssl, output, query, sources):
+          verify_ssl, samples, output, frame_style, query, sources):
+
+    samples = samples if samples > 0 else None
 
     if not query.lower().startswith('select'):
         table_names = ', '.join(['table{}'.format(index)
@@ -331,19 +367,22 @@ def query(input_encoding, output_encoding, input_locale, output_locale,
         query = 'SELECT * FROM {} WHERE {}'.format(table_names, query)
 
     if len(sources) == 1:
-        source = detect_source(sources[0], verify_ssl=verify_ssl)
+        source = detect_source(sources[0], verify_ssl=verify_ssl, progress=True)
 
-        if source.plugin_name == 'sqlite':
+        if source.plugin_name in ('sqlite', 'postgresql'):
             # Optimization: query the db directly
             result = import_from_source(source,
                                         DEFAULT_INPUT_ENCODING,
-                                        query=query)
+                                        query=query,
+                                        samples=samples)
         else:
             if input_locale is not None:
                 with rows.locale_context(input_locale):
-                    table = import_from_source(source, DEFAULT_INPUT_ENCODING)
+                    table = import_from_source(source, DEFAULT_INPUT_ENCODING,
+                            samples=samples)
             else:
-                table = import_from_source(source, DEFAULT_INPUT_ENCODING)
+                table = import_from_source(source, DEFAULT_INPUT_ENCODING,
+                        samples=samples)
 
             sqlite_connection = sqlite3.Connection(':memory:')
             rows.export_to_sqlite(table,
@@ -356,11 +395,11 @@ def query(input_encoding, output_encoding, input_locale, output_locale,
         if input_locale is not None:
             with rows.locale_context(input_locale):
                 tables = [_import_table(source, encoding=input_encoding,
-                                        verify_ssl=verify_ssl)
+                                        verify_ssl=verify_ssl, samples=samples)
                           for source in sources]
         else:
             tables = [_import_table(source, encoding=input_encoding,
-                                    verify_ssl=verify_ssl)
+                                    verify_ssl=verify_ssl, samples=samples)
                       for source in sources]
 
         sqlite_connection = sqlite3.Connection(':memory:')
@@ -378,9 +417,11 @@ def query(input_encoding, output_encoding, input_locale, output_locale,
         fobj = BytesIO()
         if output_locale is not None:
             with rows.locale_context(output_locale):
-                rows.export_to_txt(result, fobj, encoding=output_encoding)
+                rows.export_to_txt(result, fobj, encoding=output_encoding,
+                                   frame_style=frame_style)
         else:
-            rows.export_to_txt(result, fobj, encoding=output_encoding)
+            rows.export_to_txt(result, fobj, encoding=output_encoding,
+                               frame_style=frame_style)
         fobj.seek(0)
         click.echo(fobj.read())
     else:
@@ -392,9 +433,9 @@ def query(input_encoding, output_encoding, input_locale, output_locale,
 
 
 @cli.command(name='schema', help='Identifies table schema')
-@click.option('--input-encoding')
+@click.option('--input-encoding', default='utf-8')
 @click.option('--input-locale')
-@click.option('--verify-ssl', default=True, type=bool)
+@click.option('--verify-ssl', type=bool, default=True)
 @click.option('-f', '--format', 'output_format', default='txt',
               type=click.Choice(('txt', 'sql',  'django')))
 @click.option('--fields',
@@ -411,7 +452,7 @@ def schema(input_encoding, input_locale, verify_ssl, output_format, fields,
     samples = samples if samples > 0 else None
     import_fields = _get_import_fields(fields, fields_exclude)
 
-    source = detect_source(source, verify_ssl=verify_ssl)
+    source = detect_source(source, verify_ssl=verify_ssl, progress=True)
     # TODO: make it lazy
     if input_locale is not None:
         with rows.locale_context(input_locale):
@@ -431,6 +472,133 @@ def schema(input_encoding, input_locale, verify_ssl, output_format, fields,
     else:
         output = open(output, mode='w', encoding='utf-8')
     rows.fields.generate_schema(table, export_fields, output_format, output)
+
+
+@cli.command(name='csv2sqlite', help='Convert one or more CSV files to SQLite')
+@click.option('--batch-size', default=10000)
+@click.option('--samples', type=int, default=5000,
+              help='Number of rows to determine the field types (0 = all)')
+@click.option('--input-encoding', default='utf-8')
+@click.option('--dialect', default=None)
+@click.option('--schemas', default=None)
+@click.argument('sources', nargs=-1, required=True)
+@click.argument('output', required=True)
+def command_csv2sqlite(batch_size, samples, input_encoding, dialect, schemas,
+                       sources, output):
+
+    inputs = [pathlib.Path(filename) for filename in sources]
+    output = pathlib.Path(output)
+    table_names = make_header([filename.name.split('.')[0]
+                               for filename in inputs])
+    schemas = _get_schemas_for_inputs(schemas, inputs)
+
+    for filename, table_name, schema in zip(inputs, table_names, schemas):
+        prefix = '[{filename} -> {db_filename}#{tablename}]'.format(
+            db_filename=output.name,
+            tablename=table_name,
+            filename=filename.name,
+        )
+        pre_prefix = '{} (detecting data types)'.format(prefix)
+        progress = ProgressBar(prefix=prefix, pre_prefix=pre_prefix)
+        csv2sqlite(
+            six.text_type(filename),
+            six.text_type(output),
+            dialect=dialect,
+            table_name=table_name,
+            samples=samples,
+            batch_size=batch_size,
+            callback=progress.update,
+            encoding=input_encoding,
+            schema=schema,
+        )
+        progress.close()
+
+
+@cli.command(name='sqlite2csv', help='Convert a SQLite table into CSV')
+@click.option('--batch-size', default=10000)
+@click.option('--dialect', default='excel')
+@click.argument('source', required=True)
+@click.argument('table_name', required=True)
+@click.argument('output', required=True)
+def command_sqlite2csv(batch_size, dialect, source, table_name, output):
+
+    input_filename = pathlib.Path(source)
+    output_filename = pathlib.Path(output)
+    prefix = '[{db_filename}#{tablename} -> {filename}]'.format(
+        db_filename=input_filename.name,
+        tablename=table_name,
+        filename=output_filename.name,
+    )
+    progress = ProgressBar(prefix=prefix, pre_prefix='')
+    sqlite2csv(
+        input_filename=six.text_type(input_filename),
+        table_name=table_name,
+        dialect=dialect,
+        output_filename=six.text_type(output_filename),
+        batch_size=batch_size,
+        callback=progress.update,
+    )
+    progress.close()
+
+
+@cli.command(name='pgimport', help='Import a CSV file into a PostgreSQL table')
+@click.option('--input-encoding', default='utf-8')
+@click.option('--no-create-table', type=bool, default=False)
+@click.option('--dialect', default=None)
+@click.option('--schema', default=None)
+@click.argument('source', required=True)
+@click.argument('database_uri', required=True)
+@click.argument('table_name', required=True)
+def command_pgimport(input_encoding, no_create_table, dialect, schema, source,
+                     database_uri, table_name):
+
+    progress = ProgressBar(
+        prefix='Importing data',
+        pre_prefix='Detecting file size',
+        unit='bytes',
+    )
+    try:
+        total_size = uncompressed_size(source)
+    except (RuntimeError, ValueError):
+        total_size = None
+    else:
+        progress.total = total_size
+    progress.description = 'Analyzing source file'
+    schemas = _get_schemas_for_inputs([schema] if schema else None, [source])
+    import_meta = pgimport(
+        filename=source,
+        encoding=input_encoding,
+        dialect=dialect,
+        database_uri=database_uri,
+        create_table=not no_create_table,
+        table_name=table_name,
+        callback=progress.update,
+        schema=schemas[0],
+    )
+    progress.description = \
+        '{} rows imported'.format(import_meta['rows_imported'])
+    progress.close()
+
+
+@cli.command(name='pgexport', help='Export a PostgreSQL table into a CSV file')
+@click.option('--output-encoding', default='utf-8')
+@click.option('--dialect', default='excel')
+@click.argument('database_uri', required=True)
+@click.argument('table_name', required=True)
+@click.argument('destination', required=True)
+def command_pgexport(output_encoding, dialect, database_uri, table_name,
+                     destination):
+
+    updater = ProgressBar(prefix='Exporting data', unit='bytes')
+    pgexport(
+        database_uri=database_uri,
+        table_name=table_name,
+        filename=destination,
+        encoding=output_encoding,
+        dialect=dialect,
+        callback=updater.update,
+    )
+    updater.close()
 
 
 if __name__ == '__main__':