+++ rows/plugins/plugin_csv.py
@@ -27,34 +27,39 @@ from rows.plugins.utils import (create_table, get_filename_and_fobj,
 
 sniffer = unicodecsv.Sniffer()
 
+def fix_dialect(dialect):
+    if not dialect.doublequote and dialect.escapechar is None:
+        dialect.doublequote = True
+
+    if dialect.quoting == unicodecsv.QUOTE_MINIMAL and dialect.quotechar == "'":
+        # Python csv's Sniffer seems to detect a wrong quotechar when
+        # quoting is minimal
+        dialect.quotechar = '"'
+
 if six.PY2:
 
     def discover_dialect(sample, encoding=None,
                          delimiters=(b',', b';', b'\t', b'|')):
-        """Discover a CSV dialect based on a sample size
+        """Discover a CSV dialect based on a sample size.
 
         `encoding` is not used (Python 2)
         """
-
         try:
             dialect = sniffer.sniff(sample, delimiters=delimiters)
 
-        except unicodecsv.Error: # Couldn't detect: fall back to 'excel'
+        except unicodecsv.Error:  # Couldn't detect: fall back to 'excel'
             dialect = unicodecsv.excel
 
-        if not dialect.doublequote and dialect.escapechar is None:
-            dialect.doublequote = True
-
+        fix_dialect(dialect)
         return dialect
 
 elif six.PY3:
 
     def discover_dialect(sample, encoding, delimiters=(',', ';', '\t', '|')):
-        """Discover a CSV dialect based on a sample size
+        """Discover a CSV dialect based on a sample size.
 
         `sample` must be `bytes` and an `encoding must be provided (Python 3)
         """
-
         # `csv.Sniffer.sniff` on Python 3 requires a `str` object. If we take a
         # sample from the `bytes` object and it happens to end in the middle of
         # a character which has more than one byte, we're going to have an
@@ -77,18 +82,15 @@ elif six.PY3:
         try:
             dialect = sniffer.sniff(decoded, delimiters=delimiters)
 
-        except unicodecsv.Error: # Couldn't detect: fall back to 'excel'
+        except unicodecsv.Error:  # Couldn't detect: fall back to 'excel'
             dialect = unicodecsv.excel
 
-        if not dialect.doublequote and dialect.escapechar is None:
-            dialect.doublequote = True
-
+        fix_dialect(dialect)
         return dialect
 
 
 def read_sample(fobj, sample):
-    "Read `sample` bytes from `fobj` and return the cursor to where it was"
-
+    """Read `sample` bytes from `fobj` and return the cursor to where it was."""
     cursor = fobj.tell()
     data = fobj.read(sample)
     fobj.seek(cursor)
@@ -97,12 +99,11 @@ def read_sample(fobj, sample):
 
 def import_from_csv(filename_or_fobj, encoding='utf-8', dialect=None,
                     sample_size=262144, *args, **kwargs):
-    """Import data from a CSV file (automatically detects dialect)
+    """Import data from a CSV file (automatically detects dialect).
 
     If a file-like object is provided it MUST be in binary mode, like in
     `open(filename, mode='rb')`.
     """
-
     filename, fobj = get_filename_and_fobj(filename_or_fobj, mode='rb')
 
     if dialect is None:
@@ -113,21 +114,21 @@ def import_from_csv(filename_or_fobj, encoding='utf-8', dialect=None,
 
     meta = {'imported_from': 'csv',
             'filename': filename,
-            'encoding': encoding,}
+            'encoding': encoding}
     return create_table(reader, meta=meta, *args, **kwargs)
 
 
 def export_to_csv(table, filename_or_fobj=None, encoding='utf-8',
                   dialect=unicodecsv.excel, batch_size=100, callback=None,
                   *args, **kwargs):
-    """Export a `rows.Table` to a CSV file
+    """Export a `rows.Table` to a CSV file.
+
 
     If a file-like object is provided it MUST be in binary mode, like in
     `open(filename, mode='wb')`.
     If not filename/fobj is provided, the function returns a string with CSV
     contents.
     """
-
     # TODO: will work only if table.fields is OrderedDict
     # TODO: should use fobj? What about creating a method like json.dumps?