+++ rows/plugins/sqlite.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 
-# Copyright 2014-2017 Álvaro Justen <https://github.com/turicas/rows/>
+# Copyright 2014-2018 Álvaro Justen <https://github.com/turicas/rows/>
 
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Lesser General Public License as published by
@@ -17,6 +17,7 @@
 
 from __future__ import unicode_literals
 
+import datetime
 import sqlite3
 import string
 
@@ -51,13 +52,21 @@ def _python_to_sqlite(field_types):
         if field_type in (
                 fields.BinaryField,
                 fields.BoolField,
-                fields.DateField,
-                fields.DatetimeField,
                 fields.FloatField,
                 fields.IntegerField,
                 fields.TextField
         ):
             return value
+
+        elif field_type in (fields.DateField, fields.DatetimeField):
+            if value is None:
+                return None
+            elif isinstance(value, (datetime.date, datetime.datetime)):
+                return value.isoformat()
+            elif isinstance(value, (six.binary_type, six.text_type)):
+                return value
+            else:
+                raise ValueError('Cannot serialize date value: {}'.format(repr(value)))
 
         elif field_type in (fields.DecimalField,
                             fields.PercentField):
@@ -83,14 +92,13 @@ def _get_connection(filename_or_connection):
 
 
 def _valid_table_name(name):
-    '''Verify if a given table name is valid for `rows`
+    """Verify if a given table name is valid for `rows`.
 
     Rules:
     - Should start with a letter or '_'
     - Letters can be capitalized or not
     - Acceps letters, numbers and _
-    '''
-
+    """
     if name[0] not in '_' + string.ascii_letters or \
        not set(name).issubset('_' + string.ascii_letters + string.digits):
         return False
@@ -101,7 +109,7 @@ def _valid_table_name(name):
 
 def import_from_sqlite(filename_or_connection, table_name='table1', query=None,
                        query_args=None, *args, **kwargs):
-
+    """Return a rows.Table with data from SQLite database."""
     connection = _get_connection(filename_or_connection)
     cursor = connection.cursor()
 
@@ -127,7 +135,6 @@ def export_to_sqlite(table, filename_or_connection, table_name=None,
                      table_name_format='table{index}', batch_size=100,
                      callback=None, *args, **kwargs):
     # TODO: should add transaction support?
-
     prepared_table = prepare_to_export(table, *args, **kwargs)
     connection = _get_connection(filename_or_connection)
     cursor = connection.cursor()
@@ -161,11 +168,12 @@ def export_to_sqlite(table, filename_or_connection, table_name=None,
             cursor.executemany(insert_sql, map(_convert_row, batch))
 
     else:
-        total = 0
+        total_written = 0
         for batch in ipartition(prepared_table, batch_size):
             cursor.executemany(insert_sql, map(_convert_row, batch))
-            total += len(batch)
-            callback(total)
+            written = len(batch)
+            total_written += written
+            callback(written, total_written)
 
     connection.commit()
     return connection