+++ rows/table.py
@@ -18,10 +18,14 @@
 from __future__ import unicode_literals
 
 import os
-from collections import MutableSequence, OrderedDict, Sized, namedtuple
+from collections import OrderedDict, namedtuple
 from operator import itemgetter
 
 import six
+if six.PY2:
+    from collections import MutableSequence, Sized
+elif six.PY3:
+    from collections.abc import MutableSequence, Sized
 
 
 class Table(MutableSequence):