+++ tests/tests_plugin_xlsx.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 
-# Copyright 2014-2017 Álvaro Justen <https://github.com/turicas/rows/>
+# Copyright 2014-2018 Álvaro Justen <https://github.com/turicas/rows/>
 
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Lesser General Public License as published by
@@ -17,9 +17,11 @@
 
 from __future__ import unicode_literals
 
+import datetime
 import tempfile
 import unittest
 from collections import OrderedDict
+from decimal import Decimal
 from io import BytesIO
 
 import mock
@@ -140,3 +142,41 @@ class PluginXlsxTestCase(utils.RowsTestMixIn, unittest.TestCase):
 
         table2 = rows.import_from_xlsx(filename)
         self.assert_table_equal(table, table2)
+
+    @mock.patch('rows.plugins.xlsx.create_table')
+    def test_start_and_end_row(self, mocked_create_table):
+        rows.import_from_xlsx(
+            self.filename,
+            start_row=3, end_row=5,
+            start_column=2, end_column=5,
+        )
+        self.assertTrue(mocked_create_table.called)
+        self.assertEqual(mocked_create_table.call_count, 1)
+        call_args = mocked_create_table.call_args_list[0]
+        expected_data = [
+            [4.56, 4.56, '12%', datetime.datetime(2050, 1, 2, 0, 0)],
+            [7.89, 7.89, '13.64%', datetime.datetime(2015, 8, 18, 0, 0)],
+            [9.87, 9.87, '13.14%', datetime.datetime(2015, 3, 4, 0, 0)],
+        ]
+        self.assertEqual(expected_data, call_args[0][0])
+
+    def test_issue_290_can_read_sheet(self):
+        result = rows.import_from_xlsx('tests/data/text_in_percent_cell.xlsx')
+        # Before fixing the first part of #290, this would simply crash
+        assert True
+
+    def test_issue_290_one_hundred_read_as_1(self):
+        result = rows.import_from_xlsx('tests/data/text_in_percent_cell.xlsx')
+        # As this test is written, file numeric file contents on first column are
+        # 100%, 23.20%, 1.00%, 10.00%, 100.00%
+        assert result[0][0] == Decimal('1')
+        assert result[1][0] == Decimal('0.2320')
+        assert result[2][0] == Decimal('0.01')
+        assert result[3][0] == Decimal('0.1')
+        assert result[4][0] == Decimal('1')
+
+    def test_issue_290_textual_value_in_percent_col_is_preserved(self):
+        result = rows.import_from_xlsx('tests/data/text_in_percent_cell.xlsx')
+        # As this test is written, file contents on first column are
+        # 100%, 23.20%, 1.00%, 10.00%, 100.00%
+        assert result[5][1] == 'text'