+++ docs/changelog.md
@@ -26,6 +26,27 @@
   (CSV and SQLite plugins)
 - [#234](https://github.com/turicas/rows/issues/234) Remove `BinaryField` from
   the default list of detection types
+- [#270](https://github.com/turicas/rows/pull/270) Added options to export
+  pretty text table frames (TXT plugin)
+- Add `csv2sqlite` CLI subcommand
+- Create `rows.utils.open_compressed` helper function: can read/write files,
+  automatically dealing with on-the-fly compression
+- [#274](https://github.com/turicas/rows/issues/274) `start_row` and
+  `start_column` now behave the same way in XLS and XLSX
+- [#261](https://github.com/turicas/rows/issues/261) Add support to `end_row`
+  and `end_column` on XLS and XLSX (thanks
+  [@Lrcezimbra](https://github.com/Lrcezimbra) for the suggestion)
+- Add progress bar support to `utils.download_file`
+- Enable progress bar when downloading from CLI
+- Add helper class CsvLazyDictWriter
+- Add pgimport and pgexport functions and CLI commands
+- Add sqlite2csv function and CLI command
+- [#4](https://github.com/turicas/rows/issues/4) Add PostgreSQL plugin (thanks
+  to [@juliano777](https://github.com/juliano777))
+- [#290](https://github.com/turicas/rows/pull/290) Fix percent formatting
+  reading on XLSX and ODS file formats (thanks to
+  [@jsbueno](https://github.com/jsbueno))
+
 
 ### Bug Fixes